Separate tests for backwards compatibility into their own suite
