+++ openverse-api/test/v1_integration_test.py
@@ -56,39 +56,6 @@ def test_link_shortener_resolve(link_shortener_fixture):
     assert response.status_code == 301
 
 
-def test_old_stats_endpoint():
-    response = requests.get(
-        f'{API_URL}/v1/sources?type=images',
-        allow_redirects=False,
-        verify=False
-    )
-    assert response.status_code == 301
-    assert response.is_permanent_redirect
-    assert response.headers.get('Location') == '/v1/images/stats'
-
-
-def test_old_related_images_endpoint():
-    response = requests.get(
-        f'{API_URL}/v1/recommendations/images/xyz',
-        allow_redirects=False,
-        verify=False
-    )
-    assert response.status_code == 301
-    assert response.is_permanent_redirect
-    assert response.headers.get('Location') == '/v1/images/xyz/recommendations'
-
-
-def test_old_thumbs_endpoint():
-    response = requests.get(
-        f'{API_URL}/v1/thumbs/xyz',
-        allow_redirects=False,
-        verify=False
-    )
-    assert response.status_code == 301
-    assert response.is_permanent_redirect
-    assert response.headers.get('Location') == '/v1/images/xyz/thumb'
-
-
 @pytest.mark.skip(reason="Disabled feature")
 @pytest.fixture
 def test_list_create(image_fixture):
